Ask for user guidance when unpushed commits are about to be deployed
A prompt for user guidance comes up when there are commits to be
deployed, which have not been pushed to remote master.

Up until now, the prompt was thrown when the head of local
branch was different from remote master's. As a result, user was asked
even if the changes were pushed to remote master but were not the
latest ones.